% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMetrics_merged_NDRset.R
\docType{data}
\name{AllMetrics_merged_NDRset}
\alias{AllMetrics_merged_NDRset}
\title{All associated conventional and alternative response metrics for 339 cell lines and 265 drugs}
\format{A data frame with 75559 rows and 34 variables:
\describe{
  \item{CellLine}{Name of the cell line}
  \item{Drug}{Name of the drug, can be also Drug ID.}
  \item{IC50/GR50/NDR50/PI50}{Half maximal inhibitory concentrations}
  \item{LN_IC50/LN_GR50/LN_NDR50/LN_PI50}{log Half maximal inhibitory concentrations}
  \item{AUC/GRAOC/NDR_AOC/PI_AOC}{Area Under/Over the Curves}
  \item{EC50/GEC50/NDR_EC50/PI_EC50}{Half maximal effective concentrations}
  \item{Emax/GRmax/NDRmax/PImax}{Maximal response}
  \item{IC_inf/GR_inf/NDR_inf/PI_inf}{Response at infinite concentrations}
  \item{IC_HillCoeff,GR_HillCoeff,NDR_HillCoeff,PI_HillCoeff}{Slope of the dose response curves}
  \item{IC_R2,GR_R2,NDR_R2,PI_R2}{Goodness of fit of sigmoidal curves}
  }}
\usage{
AllMetrics_merged_NDRset
}
\description{
All associated conventional metrics:IC50,EC50,Emax,ICHill,IC_inf,IC_R2,AUC
All associated alternative metrics:GR50,GEC50,GRmax,GRHill,GR_inf,GR_R2,GRAOC; Similar metrices calculated for NDR and PI as well
}
\details{
A dataframe with all associated metrics calculated by AllMetrics_Calc and ICMetrics_Calc function
}
\keyword{datasets}
