% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_GDSC.R
\name{calculate_concentrations}
\alias{calculate_concentrations}
\title{Calculate_concentrations}
\usage{
calculate_concentrations(RawData, writeFile = F,
  outFile = "AllCellLines_Drugs_Conc.txt")
}
\arguments{
\item{RawData}{read  \code{\link{AllDrugs_CellLines_GDSC}} for data format.}

\item{writeFile}{(default FALSE) write results into file}

\item{outFile}{destination of results}
}
\value{
data.frame with CellLine, Drug and corresponding Concentration values.
}
\description{
Calculate concentrations based on fold dilution and Max. conc.
}
