% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellLines_var.R
\name{CellLines_var}
\alias{CellLines_var}
\title{Calculate variance of GR and Rel_via for each cell line}
\usage{
CellLines_var(Gemcitabine_conc_filt, metric = c("GR", "Rel_via"))
}
\arguments{
\item{Gemcitabine_conc_filt}{Gemcitabine responses (GR/Rel_via)}

\item{metric}{GR/Rel_via}
}
\description{
Only responses corresponding to concentrations<32Mm considered
}
