% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conventional_GR_allCalculatedMetrics_merged.R
\docType{data}
\name{Conventional_GR_allCalculatedMetrics_merged}
\alias{Conventional_GR_allCalculatedMetrics_merged}
\title{All associated conventional and GR metrics for 173 cell lines and 265 drugs}
\format{A data frame with 38393 rows and 19 variables:
\describe{
  \item{CellLine}{Name of the cell line}
  \item{Drug}{Name of the drug, can be also Drug ID.}
  \item{IC50/GR50}{Half maximal inhibitory concentrations}
  \item{LN_IC50/LN_GR50}{log Half maximal inhibitory concentrations}
  \item{AUC/AAC/GRAOC}{Area Under/Above/Over the Curves}
  \item{EC50/GEC50}{Half maximal effective concentrations}
  \item{Emax/GRmax}{Maximal response}
  \item{IC_inf/GR_inf}{Response at infinite concentrations}
  \item{IC_HillCoeff,GR_HillCoeff}{Slope of the dose response curves}
  \item{IC_R2,GR_R2}{Goodness of fit of sigmoidal curves}
  }}
\usage{
Conventional_GR_allCalculatedMetrics_merged
}
\description{
All associated conventional metrics:IC50,EC50,Emax,ICHill,IC_inf,IC_R2,AUC
All associated alternative metrics:GR50,GEC50,GRmax,GRHill,GR_inf,GR_R2,GRAOC
}
\details{
A dataframe with all associated metrics calculated by AllMetrics_Calc and ICMetrics_Calc function
}
\keyword{datasets}
