% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_ID_all.R
\docType{data}
\name{COSMIC_ID_all}
\alias{COSMIC_ID_all}
\title{COSMIC ids of 1065 cell lines}
\format{A data frame with 1065 rows and 2 variables:
\describe{
  \item{CellLine}{Name of the cell line}
  \item{Drug}{Name of the drug, can be also Drug ID.}
  }}
\source{
\url{https://www.cancerrxgene.org/downloads}
}
\usage{
COSMIC_ID_all
}
\description{
Cell line annotations downloaded from GDSC \url{https://www.cancerrxgene.org/downloads}
}
\details{
A dataframe with COSMIC ids and the corresponding cell line names
}
\keyword{datasets}
