% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSC_orig_IC50_AUC.R
\docType{data}
\name{GDSC_orig_IC50_AUC}
\alias{GDSC_orig_IC50_AUC}
\title{Original IC50s and AUCs of GDSC data}
\format{A data frame with 224202 rows and 4 variables:
\describe{
  \item{CellLine}{Name of the cell line}
  \item{Drug}{Name of the drug}
  \item{LN_IC50}{Half maximal Inhibitory Concentrations}
  \item{AUC}{Area Under the Curve}
  }}
\source{
\url{https://www.cancerrxgene.org/downloads}
}
\usage{
GDSC_orig_IC50_AUC
}
\description{
Drug response information downloaded from GDSC \url{https://www.cancerrxgene.org/downloads}
IC50s and AUCs in this correspond to the values generated based on non-linear mixed effects model. For details, refer \url{https://www.ncbi.nlm.nih.gov/pubmed/27180993}
}
\details{
A dataframe with cell lines, drugs, LNIC50s and AUCs
}
\keyword{datasets}
