% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSCTools_Features_prep.R
\name{GDSCTools_Features_prep}
\alias{GDSCTools_Features_prep}
\title{Generate response matrix in GDSC format}
\usage{
GDSCTools_Features_prep(GDSC_IC50_GR50, COSMIC_ID, metric, dataset)
}
\arguments{
\item{metric}{IC50/AUC/Emax/GR50/GRAOC/GRmax/NDR50/NDRAOC/PI50/PIAOC}
}
\value{
A dataframe with drugs as columns and cell lines as rows; Drug response metrics values are filled up in matrices
}
\description{
Generate response matrix in GDSC format
}
