% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICMetrics_Calc.R
\name{ICMetrics_Calc}
\alias{ICMetrics_Calc}
\title{Calculate the conventional metrics}
\usage{
ICMetrics_Calc(All_GDSC_IC_data, writeFile = T, fileName)
}
\arguments{
\item{All_GDSC_IC_data}{A data frame with CellLine, Drug, Concentration and Relative Viability of all GDSC data}
}
\value{
A data frame with all the conventional metrics
}
\description{
Fit sigmoid curves to compute the conventional metrics (IC50,EC50,Emax,ICinf,IChill,AAC,ICR2) for GDSC
Significance of curve fits assessed based on the conditions specified in Marc Hafner's paper
Concentrations are capped to the highest concentration value, when the GEC50s are 10 times more than the highest concentration; Same also applies, when the p-values resulting from the significance of fits is >0.5
}
