% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDR_PI_GR_metrics.R
\name{NDR_PI_GRmetrics}
\alias{NDR_PI_GRmetrics}
\title{Calculate NDR, PI and GR values based on the approach described in Gupta et al., 2018}
\usage{
NDR_PI_GRmetrics(GDSC_rawData, CellDivisionTimes, writeFile = T,
  outFile = "Calculated_Values/AllCellLines_Drugs_RelVia_NDR_GR_PI.txt")
}
\arguments{
\item{GDSC_rawData}{Raw data file with information about fold dilutions, maximum concentration, viabilities, positive and negative control}

\item{Cell_division_times}{Cell Division times extracted from GCSI data}
}
\value{
A dataframe with the cell line, drug, concentration, relative viability, NDR, PI and GR metrics
}
\description{
Negative control, positive control and viabilities at starting (0h) and end points (72h) of the assay are needed for NDR calculations. This information is not available in GDSC.
Therefore, the End points (Measurements after 72h) and Doubling times (GCSI) are used to estimate the starting point values values (Negative control,NC(t)=NC0*(2^t/DT); NC0=NCt/(2^t/DT))
}
