% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orig_Target.R
\docType{data}
\name{Orig_Target}
\alias{Orig_Target}
\title{Primary targets of GDSC drugs}
\format{A data frame with 1711 rows and 2 variables:
\describe{
  \item{Drug}{Name of the drug}
  \item{Target}{Primary target of the drug}
  }}
\source{
\url{https://www.cancerrxgene.org/downloads}

\url{https://www.ebi.ac.uk/chembl/}
}
\usage{
Orig_Target
}
\description{
Target information downloaded from GDSC \url{https://www.cancerrxgene.org/downloads} and ChEMBL 23 \url{https://www.ebi.ac.uk/chembl/}
Only those bioactivities with target confidence of 7 and above were retained; Further the bioactivities were filtered based on the following criteria:
Kd/Ki/IC50 < 100nM; Inhibition % > 80% and Residual activities < 20%
}
\details{
A dataframe with drugs and the corresponding primary targets
}
\keyword{datasets}
