% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_GDSC.R
\name{preprocess_GDSC}
\alias{preprocess_GDSC}
\title{Preprocess GDSc data}
\usage{
preprocess_GDSC(RawData, writeFile = T,
  outFile = "Calculated_Values/AllCellLines_Drugs_Conc_Via.txt")
}
\arguments{
\item{RawData}{read \code{\link{AllDrugs_CellLines_GDSC}}  for data format}

\item{writeFile}{(default TRUE) write results into file}

\item{outFile}{destination of results}
}
\value{
data.frame with CellLine, Drug and corresponding Concentration and cell viability values
}
\description{
Preprocess GDSc data
}
