% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_coincidence.R
\name{Random_coincidence}
\alias{Random_coincidence}
\title{Simulate overlap coincidence between the different response metrics}
\usage{
Random_coincidence(ANOVA_res)
}
\arguments{
\item{ANOVA_res}{ANOVA results of one of the metrics}
}
\value{
A data frame with the number of associations and the random overlap percentages
}
\description{
Estimate the probability of finding overlaps between 2 different metrics after random sampling of a set of gene drug associations;
Different association thresholds are used
}
