% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmoidFit_ComputeR2_IC50.R
\name{SigmoidFit_ComputeR2_IC50}
\alias{SigmoidFit_ComputeR2_IC50}
\title{Fit Sigmoid curve; Compute R2 & IC50}
\usage{
SigmoidFit_ComputeR2_IC50(Gemcitabine_conc_filt, writeFile = T,
  outFile = "Calculated_Values/Gemcitabine_GDSC_FittedSigmoid_R2_IC50.txt")
}
\arguments{
\item{Gemcitabine_conc_filt}{A data frame with concentrations and Relative viabilities (Concentration ranges: < 32Mm)}
}
\value{
A data frame with cell lines together with R2 and IC50 extracted from curve fitting
}
\description{
Fit Sigmoid curve; Compute R2 & IC50
}
