% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_R_Drug_AllMetrics.R
\docType{data}
\name{Test_R_Drug_AllMetrics}
\alias{Test_R_Drug_AllMetrics}
\title{Test set predictions of drug response models}
\format{A data frame with 5 rows and 4 variables:
\describe{
  \item{GRAOC/AUC/GR50/IC50}{Pearson correlation between the observed and predicted values}
  }}
\source{
\url{https://www.cancerrxgene.org/downloads}
}
\usage{
Test_R_Drug_AllMetrics
}
\description{
Test set predicts correspond to the 5 different test sets generated by random splitting; Each test set constitutes 20% of the data
}
\details{
A dataframe with the predictions from models based on GRAOC, AUC, GR50 and IC50
}
\keyword{datasets}
